templates/api.c:	Fixed wrong type for gliss_run_and_count()

+++ templates/api.c
@@ -692,7 +692,7 @@ $(end)
  * @param	sim	the simulator which we simulate within
  * @return number of executed instructions
  * */
-unsigned int $(proc)_run_and_count_inst($(proc)_sim_t *sim)
+uint64_t $(proc)_run_and_count_inst($(proc)_sim_t *sim)
 {
 	unsigned int i = 0;
     uint32_t num_bloc;
@@ -868,7 +868,7 @@ $(end)
  * @param	sim	the simulator which we simulate within
  * @return number of executed instructions
  * */
-unsigned int $(proc)_run_and_count_inst($(proc)_sim_t *sim)
+uint64_t $(proc)_run_and_count_inst($(proc)_sim_t *sim)
 {
 	unsigned int i = 0;
     $(proc)_state_t*   state     = sim->state;
